# miraTP2/openapi.yaml
openapi: 3.0.3
info:
  title: Mira Notes API
  description: TP#2 - API /api/v1/notes du projet fil rouge Mira.
  version: "1.0.0"
servers:
  - url: http://localhost:8080

paths:
  /api/v1/notes:
    post:
      summary: Créer une note
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Note créée
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      summary: Lister les notes
      operationId: listNotes
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/StatusFilter"
      responses:
        "200":
          description: Liste des notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"

  /api/v1/notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Récupérer une note
      operationId: getNote
      responses:
        "200":
          description: Note trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Mettre à jour partiellement une note
      operationId: updateNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Note mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Supprimer une note
      operationId: deleteNote
      responses:
        "204":
          description: Note supprimée
        "404":
          $ref: "#/components/responses/NotFound"

  /api/v1/search:
    get:
      summary: Recherche texte simple sur les notes
      operationId: searchNotes
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Terme recherché (insensible à la casse, sur title + content)
      responses:
        "200":
          description: Notes correspondantes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"

components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 20
        maximum: 100
      description: Nombre maximum de notes retournées (défaut 20, plafonné à 100).
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    StatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/NoteStatus"

  responses:
    BadRequest:
      description: Payload ou paramètres invalides
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Note inexistante
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

  schemas:
    NoteStatus:
      type: string
      enum: [active, archived]

    Note:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        status:
          $ref: "#/components/schemas/NoteStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, title, content, status, created_at, updated_at]

    CreateNoteRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          maxLength: 10000
        status:
          $ref: "#/components/schemas/NoteStatus"
      required: [title]

    UpdateNoteRequest:
      type: object
      description: Seuls les champs fournis sont modifiés.
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          maxLength: 10000
        status:
          $ref: "#/components/schemas/NoteStatus"

    Meta:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [total]

    NoteEnvelope:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Note"
      required: [data]

    NoteListEnvelope:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]

    ErrorBody:
      type: object
      properties:
        code:
          type: string
          enum: [invalid_json, validation_error, not_found, internal_error, timeout]
        message:
          type: string
      required: [code, message]

    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
      required: [error]